+++ RELEASENOTES.md
@@ -3,6 +3,8 @@
 
 * base code on mulle-bootstrap library functions for exekutor and logging
 * add multiple options
+* check for mulle-bootstrap version
+* rename project from mulle-install to mulle-build
 
 
 0.2